Move casement cli into its own module
